GRAPHICS, SETCOLOR, PLOT, and COLOR in Assembly
This is the start of what will be a game demo eventually. It is the  fully commented code through parts 7 and 7.2 in my YouTube Assembly Language tutorial series. There are two changes from the code in the video (that will be shown in next video). Changed .WORD START to .WORD BEGIN for the DOS runnable start address at the end of the program. And as I thought about in the video, changed CLC, ADC LUMNCE in Setcolor subroutine to ORA LUMNCE since the low byte of HUE in accumulator is empty after we shifted all the bits to the high byte and the high byte in LUMNCE is empty so it's OK to use ORA to merge the two values.